requirements.txt
asgiref==3.6.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Django==4.2
djangorestframework==3.14.0
docopt==0.6.2
Flask==2.2.2
fonttools==4.39.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.24.2
packaging==23.0
Pillow==9.5.0
protobuf==3.20.1
psycopg2==2.9.5
pyparsing==3.0.9
PySimpleGUI==4.60.4
python-dotenv==0.21.0
pytz==2022.6
qutip==4.7.1
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.9
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.6
urllib3==1.26.12
Werkzeug==2.2.2
